Add after_commit hooks to resources
Allows users to run logic after database transactions have been
committed. Follows the same logic as before_commit hooks, allowing hooks
to be restricted to specific actions & to run multiple hooks for the
same resource on the same action.
