+++ lib/graphiti.rb
@@ -61,7 +61,7 @@ require "graphiti/util/relationship_payload"
 require "graphiti/util/persistence"
 require "graphiti/util/validation_response"
 require "graphiti/util/sideload"
-require "graphiti/util/hooks"
+require "graphiti/util/transaction_hooks_recorder"
 require "graphiti/util/attribute_check"
 require "graphiti/util/serializer_attributes"
 require "graphiti/util/serializer_relationships"
